+++ app/templates/add_record.html
@@ -30,7 +30,7 @@ Add a Record
 
 			<!-- links -->
 			<p><a href="{{ url_for('directory') }}">View the updated directory.</a></p>
-			<p><a href="{{ url_for('directory') }}">Enter additional data for this well site.</a></p>
+			<p><a href="{{ url_for( 'select_record1', letters='AF' ) }}">Enter additional data for this well site.</a></p>
 			<p><a href="{{ url_for('add_record') }}">Submit another record.</a></p>
 
 		{% else %}
