+++ app/templates/database.html
@@ -26,7 +26,7 @@ BitCan Well Database
 	  <!-- links -->
 	  <p>You can <a href="{{ url_for('directory') }}">view BitCan's well directory</a>.</p>
 	  <p>You can <a href="{{ url_for('add_record') }}">add a new well location</a>.</p>
-	  <p>You can <a href="{{ url_for( 'select_record', letters='AF' ) }}">edit an existing well location</a>.</p>
+	  <p>You can <a href="{{ url_for( 'select_record1', letters='AF' ) }}">edit an existing well location</a>.</p>
 	  <p>You can <a href="{{ url_for('home') }}">return to the web app</a>.</p>
 
     </div>
