+++ app/templates/edit_or_delete.html
@@ -45,7 +45,7 @@
 		  <!-- links -->
 		  <p class="ml-4">No. <a href="{{ url_for('home') }}">Return to the index.</a></p>
 
-		  <p class="ml-4">No. <a href="{{ url_for( 'select_record', letters='AF' ) }}">Return to the selection page.</a></p>
+		  <p class="ml-4">No. <a href="{{ url_for( 'select_record1', letters='AF' ) }}">Return to the selection page.</a></p>
 
 		<!-- table showing the selected sock -->
 		<table class="table">
@@ -91,7 +91,7 @@
 		<!-- links -->
 		<p><a href="{{ url_for('database') }}">Return to index.</a></p>
 
-		<p><a href="{{ url_for( 'select_record', letters='AF' ) }}">Return to the selection page.</a></p>
+		<p><a href="{{ url_for( 'select_record1', letters='AF' ) }}">Return to the selection page.</a></p>
 
 		<!-- show flash - based on WTForms validators
 			 see https://pythonprogramming.net/flash-flask-tutorial/
