--- app/templates/select_record.html
@@ -1,85 +1,0 @@
-{% extends 'bootstrap/base.html' %}
-
-{% block styles %}
-{{ super() }}
-	<style>
-		body { background: #e8f1f9; }
-		ul { display: flex; }
-	</style>
-{% endblock %}
-
-
-{% block title %}
-Select a Record
-{% endblock %}
-
-
-{% block content %}
-
-<div class="container">
-  <div class="row pb-5">
-    <div class="col-lg-10 mx-lg-auto">
-
-      <h1 class="pt-5 pb-2">Select a Record</h1>
-
-      <p class="lead alert alert-success">Select one record to edit or delete.</p>
-
-	  <!-- links to display a group of socks, alphabetically -->
-	  <ul class="pl-0">
-		  <li class="text-center list-group-item list-group-item-action"><a href="{{ url_for( 'select_record', letters='AF' ) }}">A &ndash; E</a></li>
-		  <li class="text-center list-group-item list-group-item-action"><a href="{{ url_for( 'select_record', letters='FM' ) }}">F &ndash; L</a></li>
-		  <li class="text-center list-group-item list-group-item-action"><a href="{{ url_for( 'select_record', letters='MS' ) }}">M &ndash; R</a></li>
-		  <li class="text-center list-group-item list-group-item-action"><a href="{{ url_for( 'select_record', letters='Sa' ) }}">S &ndash; Z</a></li>
-	  </ul>
-
-	  <p class="ml-4"><a href="{{ url_for('database') }}">Return to index.</a></p>
-
-	  <!-- start form - not using WTForms here -->
-	  <form method="POST" id="form2" action="{{ url_for('edit_or_delete') }}">
-
-		<table class="table">
-			<tr>
-				<th></th>
-				<th>Directory</th>
-				<th>Shmin</th>
-				<th>Type</th>
-				<th>Temperature</th>
-				<th>Pore Pressure</th>
-				<th>id</th>
-			</tr>
-
-			<!-- here is a for-loop to give us all the socks -->
-			{% for well in wells %}
-			<tr>
-				<td class="center-align"><input type="radio" name="id" value="{{ well.rowid }}" required></td>
-				<td>{{ well.Directory }}</td>
-				<td>{{ well.Shmin }}</td>
-				<td>{{ well.Type }}</td>
-				<td>{{ well.Temperature }}</td>
-				<td>{{ well.PorePressure }}</td>
-				<td>{{ well.rowid }}</td>
-			</tr>
-			{% endfor %}
-		</table>
-
-		<div class="form-group lead alert alert-success pb-4">
-          <p>What do you want to do with the selected record?</p>
-
-          <div class="form-check form-check-inline">
-          <label class="form-check-label"><input class="form-check-input" type="radio" name="choice" value="delete" required> Delete it</label>
-          </div>
-          <div class="form-check form-check-inline">
-          <label class="form-check-label"><input class="form-check-input" type="radio" name="choice" value="edit" required> Edit it</label>
-          </div>
-        </div><!-- end form-group -->
-
-		<div class="form-group">
-	        <input class="btn btn-success" type="submit" value="Submit Choice">
-        </div>
-
-		</form>
-		</div>
-	</div>
-</div>
-
-{% endblock %}
