+++ app/templates/select_record2.html
@@ -1,0 +1,101 @@
+{% extends 'bootstrap/base.html' %}
+
+{% block styles %}
+{{ super() }}
+	<style>
+		body { background: #e8f1f9; }
+		ul { display: flex; }
+	</style>
+{% endblock %}
+
+
+{% block title %}
+Additional Well Properties
+{% endblock %}
+
+
+{% block content %}
+
+<div class="container">
+  <div class="row pb-5">
+    <div class="col-lg-10 mx-lg-auto">
+
+      <h1 class="pt-5 pb-2">Select a Record</h1>
+
+      <p class="lead alert alert-success">Select one record to edit.</p>
+
+	  <p class="ml-4"><a href="{{ url_for('database') }}">Return to index.</a></p>
+
+	  <!-- start form - not using WTForms here -->
+	  <form method="POST" id="form2" action="{{ url_for('edit_or_delete') }}">
+		<table class="table"> 
+			<tr>
+				<th></th>
+				<th>Directory</th>
+                <th>Depth</th>
+                <th>Type</th>
+                <th>Description</th>
+                <th>Additional Notes</th>
+                <th>Shmin</th>
+                <th>Shmax</th>
+                <th>Vertical Stress</th>
+                <th>Temperature</th>
+                <th>Pore Pressure</th>
+                <th>Young's Modulus</th>
+                <th>Shear Modulus</th>
+                <th>Bulk Modulus</th>
+                <th>Poisson's Ratio</th>
+                <th>Cohesive Strength</th>
+                <th>Friction Angle</th>
+                <th>P-Wave</th>
+                <th>S-Wave</th>
+			</tr>
+
+			<!-- here is a for-loop to give us all the socks -->
+			{% for well in wells %}
+			<tr>
+				<td class="center-align"><input type="radio" name="id" value="{{ well.rowid }}" required></td>
+				<td>{{ well.Directory }}</td>
+                <td>{{ well.Depth }}</td>
+                <td>{{ well.Type }}</td>
+                <td>{{ well.AdditionalNotes }}</td>
+                <td>{{ well.Shmin }}</td>
+                <td>{{ well.Shmax }}</td>
+                <td>{{ well.VerticalStress }}</td>
+                <td>{{ well.Temperature }}</td>
+                <td>{{ well.PorePressure }}</td>
+                <td>{{ well.YoungsModulus }}</td>
+                <td>{{ well.ShearModulus }}</td>
+                <td>{{ well.BulkModulus }}</td>
+                <td>{{ well.PoissonsRatio }}</td>
+                <td>{{ well.CohesiveStrength }}</td>
+                <td>{{ well.FrictionAngle }}</td>
+                <td>{{ well.Pwave }}</td>
+                <td>{{ well.Swave }}</td>
+
+			</tr>
+			{% endfor %}
+		</table>
+
+		<div class="form-group lead alert alert-success pb-4">
+          <p>Is this the record you would like to edit?</p>
+
+          <div class="form-check form-check-inline">
+          <label class="form-check-label"><input class="form-check-input" type="radio" name="choice" value="delete" required> Delete it</label>
+          </div>
+          <div class="form-check form-check-inline">
+          <label class="form-check-label"><input class="form-check-input" type="radio" name="choice" value="edit" required> Edit it</label>
+          </div>
+        </div>
+        <!-- end form-group -->
+
+		<div class="form-group">
+	        <input class="btn btn-success" type="submit" value="Submit Choice">
+        </div>
+
+		</form>
+		</div>
+	</div>
+</div>
+
+{% endblock %}
